style: remove text < 0.

+++ app/src/main/java/com/github/moko256/twitlatte/drawable/PercentBarBackgroundDrawable.kt
@@ -25,7 +25,7 @@ import android.graphics.drawable.Drawable
  * @author moko256
  */
 class PercentBarBackgroundDrawable: Drawable() {
-    var percent: Float = 100f //0..100
+    var percent: Int = 100 //0..100
     var lineSize = 4f
     var color = 0x000000
